requirements.txt
playwright
python-dotenv
requests
google-api-python-client
google-auth
